+++ aldrin/src/low_level.rs
@@ -8,6 +8,7 @@ mod event;
 mod pending_reply;
 mod promise;
 mod proxy;
+mod reply;
 mod service;
 mod service_info;
 #[cfg(test)]
@@ -25,5 +26,6 @@ pub use event::Event;
 pub use pending_reply::PendingReply;
 pub use promise::Promise;
 pub use proxy::Proxy;
+pub use reply::Reply;
 pub use service::Service;
 pub use service_info::ServiceInfo;